initial commit of burp scanner talk
this will be a slightly expanded version of a talk i gave at SPW 2017.
there will be less live demoing and more covering of actual issues +
open source alternatives to Burp Suite

the SPW 2017 talk is at git@bitbucket.org:leejo/burp_scanner.git - but
you can't see that as it's a private repo. i'll be adapting the code
contained in that repo for reuse here